requirements.txt
fastapi[standard]
pydantic
python-multipart
redis
motor
pydantic-settings
PyJWT
passlib
cryptography
jwt
jose
python-jose
locust
slowapi==0.1.9
bcrypt==4.0.1
pymongo
bson
motor
slowapi
aiohttp
google-cloud-storage
langchain 
google-generativeai
langchain-google-genai
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0